Updated Logging Feature
Removed full server log as this could cause sync problems during high
server loads when multiple threads tried to write to it at once. I've
commented the code out instead of removing it, so in the future I may
add a .lock to the file to that only one thread would be able to write
to it at once.